lib: move gobuild, modproxy, and revproxy under lib
