+++ cmd/go-cache-plugin/setup.go
@@ -27,10 +27,10 @@ import (
 	"github.com/creachadair/taskgroup"
 	"github.com/creachadair/tlsutil"
 	"github.com/goproxy/goproxy"
-	"github.com/tailscale/go-cache-plugin/gobuild"
+	"github.com/tailscale/go-cache-plugin/lib/gobuild"
+	"github.com/tailscale/go-cache-plugin/lib/modproxy"
+	"github.com/tailscale/go-cache-plugin/lib/revproxy"
 	"github.com/tailscale/go-cache-plugin/lib/s3util"
-	"github.com/tailscale/go-cache-plugin/modproxy"
-	"github.com/tailscale/go-cache-plugin/revproxy"
 	"tailscale.com/tsweb"
 )
